tests/wpt/tests/dom/nodes/getElementsByClassName-01.htm

<!doctype html>
<html class="a">
 <head>
  <title>document.getElementsByClassName(): simple</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body class="a">
  <div id="log"></div>
  <script> test(function() {assert_array_equals(document.getElementsByClassName("\ta\n"),
                                                [document.documentElement, document.body])}) </script>
 </body>
</html>
